User delete Exception fixed

+++ src/main/java/com/ariel/course/services/UserService.java
@@ -5,7 +5,6 @@ import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.ariel.course.entities.User;
@@ -34,10 +33,11 @@ public class UserService {
 
 	public void delete(Long id) {
 		try {
-			repository.deleteById(id);
-		}catch(EmptyResultDataAccessException e) {
-			throw new ResourceNotFoundException(id);
-		}catch(DataIntegrityViolationException e) {
+			if (!repository.existsById(id)) {
+				throw new ResourceNotFoundException(id);
+			} else
+				repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
 			throw new DatabaseException(e.getMessage());
 		}
 	}
